change short to int

+++ src/main/java/com/webank/webase/node/mgr/deploy/entity/TbConfig.java
@@ -53,7 +53,7 @@ public class TbConfig {
      *
      * @mbg.generated
      */
-    private Short configType;
+    private Integer configType;
 
     /**
      *
